pyproject.toml
[project]
name = "adpo-vertaxai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "google-adk==0.5.0",
    "google-api-python-client>=2.169.0",
    "google-auth==2.29.0",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform>=1.87.0",
    "google-cloud-firestore==2.16.0",
    "google-cloud-pubsub==2.21.0",
    "requests==2.31.0",
    "uvicorn[standard]>=0.34.0",
]
